requirements.txt
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.10.5
djangorestframework==3.6.3
gunicorn==19.7.1
olefile==0.44
static3==0.7.0
django-rest-auth==0.9.1